+++ library/core/src/main/java/com/google/android/exoplayer2/text/subrip/SubripDecoder.java
@@ -41,10 +41,12 @@ public final class SubripDecoder extends SimpleSubtitleDecoder {
 
   private static final String TAG = "SubripDecoder";
 
+  // Some SRT files don't include hours in the timecode, so we use an optional group.
   private static final String SUBRIP_TIMECODE = "(?:(\\d+):)?(\\d+):(\\d+),(\\d+)";
   private static final Pattern SUBRIP_TIMING_LINE =
       Pattern.compile("\\s*(" + SUBRIP_TIMECODE + ")\\s*-->\\s*(" + SUBRIP_TIMECODE + ")\\s*");
 
+  // NOTE: Android Studio's suggestion to simplify '\\}' is incorrect [internal: b/144480183].
   private static final Pattern SUBRIP_TAG_PATTERN = Pattern.compile("\\{\\\\.*?\\}");
   private static final String SUBRIP_ALIGNMENT_TAG = "\\{\\\\an[1-9]\\}";
 
@@ -229,7 +231,8 @@ public final class SubripDecoder extends SimpleSubtitleDecoder {
   }
 
   private static long parseTimecode(Matcher matcher, int groupOffset) {
-    long timestampMs = Long.parseLong(matcher.group(groupOffset + 1)) * 60 * 60 * 1000;
+    @Nullable String hours = matcher.group(groupOffset + 1);
+    long timestampMs = hours != null ? Long.parseLong(hours) * 60 * 60 * 1000 : 0;
     timestampMs += Long.parseLong(matcher.group(groupOffset + 2)) * 60 * 1000;
     timestampMs += Long.parseLong(matcher.group(groupOffset + 3)) * 1000;
     timestampMs += Long.parseLong(matcher.group(groupOffset + 4));